+++ src/components/ItemStatus/index.tsx
@@ -3,7 +3,7 @@ import { formatEther } from 'ethers/lib/utils'
 import React from 'react'
 import { useSaleActionManager, useSaleOrderManager } from 'state/sale/hooks'
 import { SaleAction } from 'state/sale/reducer'
-import { OrderWrapperInterface } from 'orders/orders'
+import { OrderWrapperInterface } from 'orders/types'
 import { timeDifference } from 'utils/relativeTime'
 
 export const ItemStatus = ({ loading, makeOfferModalToggle, owner, address, identifier, buyOrders, sellOrders }: { loading: boolean, makeOfferModalToggle: () => void, owner: string, address: string, identifier: string, buyOrders: OrderWrapperInterface[], sellOrders: OrderWrapperInterface[] }) => {
