+++ src/state/sale/actions.ts
@@ -1,5 +1,5 @@
 import { createAction } from '@reduxjs/toolkit'
-import { OrderWrapperInterface } from 'orders/orders'
+import { OrderWrapperInterface } from 'orders/types'
 import { SaleAction } from './reducer'
 
 export const updateAction = createAction<{ action: SaleAction | null }>('sale/updateAction')
